Apply suggestions after code review
