+++ changes/370.feature.md
@@ -3,22 +3,6 @@ in `tests_modules` and `requires`.
 
 The following check has been added for `module.def` files:
 
-- `RequiredModuleAlreadyInTestsModules`/`required-module-already-in-tests-modules`.
+- `ModuleDefRequiredModuleAlreadyInTestsModules`/`module-def-required-module-already-in-tests-modules`.
 
 The check is activated for SonarQube.
-
-```module.def
-example_module_tests 1
-
-requires
-    example_module
-end
-
-tests_modules
-    example_module
-end
-```
-
-Since `example_module` is already included in `tests_modules`,
-`example_module_tests` implicitly requires `example_module` already so it is
-redundant to add it to `requires` again.
