+++ magik-checks/src/main/resources/nl/ramsolutions/sw/sonar/l10n/moduledef/rules/ModuleDefRequiredModuleAlreadyInTestsModules.html
@@ -1,4 +1,4 @@
-<p>Ensure that no module listed in <code>tests_modules</code> also appears in <code>requires</code>. 
+<p>Ensure that no module listed in <code>tests_modules</code> also appears in <code>requires</code>.
 If a module is listed as <code>tests_modules</code>, it is already required implicitly. Adding it to <code>requires</code> is redundant.</p>
 
 <h2>Non-compliant Code Example</h2>
