+++ sonar-magik-plugin/src/main/resources/nl/ramsolutions/sw/sonar/l10n/product_module_def/rules/Sonar_way_profile.json
@@ -3,8 +3,8 @@
   "ruleKeys": [
     "ModuleDefMissingDescription",
     "ModuleDefNameDoesNotMatchDirectoryName",
+    "ModhleDefRequiredModuleAlreadyInTestsModules",
     "ModuleDefSyntaxError",
-    "RequiredModuleAlreadyInTestsModules",
     "ProductDefMissingDescription",
     "ProductDefMissingTitle",
     "ProductDefNameDoesNotMatchDirectoryName",
